# docs/swagger.yaml
definitions:
...
  handlers.Room:
    properties:
      capacity:
        type: integer
      id:
        type: string
      name:
        type: string
      price:
        type: integer
      type:
        type: string
    type: object
...
  handlers.responseRoom:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.Room'
        type: array
      message:
        type: string
    type: object
info:
  contact: {}
  description: This is documentation API from Swagger. Jika ada masalah token silahkan
    coba pakai postman
  title: Swagger MeetApp By Sinau Koding API
  version: "1.0"
paths:
...
  /rooms:
    get:
      description: Retrieve all rooms in the system. Requires a valid Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.responseRoom'
      security:
      - BearerAuth: []
      summary: Get all rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a new room
      parameters:
      - description: Room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handlers.Room'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.Room'
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - rooms
  /rooms/{id}:
    delete:
      description: Delete a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - BearerAuth: []
      summary: Delete a room by ID
      tags:
      - rooms
    get:
      description: Get a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.responseRoom'
      security:
      - BearerAuth: []
      summary: Get a room by ID
      tags:
      - rooms
    put:
      consumes:
      - application/json
      description: Update a room by ID
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handlers.Room'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.responseRoom'
      security:
      - BearerAuth: []
      summary: Update a room by ID
      tags:
      - rooms
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"